Add AdoNetAppender to log events in a database
The AdoNetAppender class has been implemented for logging events to a database. This allows for easy storage and access of log events which aids in application debugging and monitoring. Configuration properties have also been added to control the connection to the database and the command to execute for logging the events.