Add DimensionEntity class and related functionalities
A new class, DimensionEntity, is introduced to manage the elements in the modeling process. A private list of DimensionEntity objects is created to store the dimension entities. A new method, ProcessEntitiesAndFillDimensions, is added to process these entities. Additionally, at every odd count of an element's run, a new DimensionEntity is created and added to the _entities list.